+++ src/js/analyzer-cdk/.npmignore
