+++ src/js/analyzer-cdk/package.json
